requirements.txt
Flask==1.1.2
google-cloud-bigquery==2.16.0
google-cloud-storage==1.38.0
pandas==1.2.4
openpyxl==3.0.7
